requirements_droidenv.txt
# platform: linux-64
# created-by: conda 25.3.0
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.2.1=pypi_0
addict=2.4.0=pypi_0
aiofiles=22.1.0=pypi_0
aiosqlite=0.21.0=pypi_0
anyio=4.9.0=pypi_0
argcomplete=3.6.1=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=3.0.0=pypi_0
attrs=25.3.0=pypi_0
babel=2.17.0=pypi_0
beautifulsoup4=4.13.3=pypi_0
blas=1.0=mkl
bleach=6.2.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.2.25=h06a4308_0
certifi=2025.1.31=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.1=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.2=pypi_0
contourpy=1.3.0=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.12.1=pypi_0
debugpy=1.8.13=pypi_0
decorator=5.2.1=pypi_0
defusedxml=0.7.1=pypi_0
deprecation=2.1.0=pypi_0
droid-backends=0.0.0=pypi_0
entrypoints=0.4=pypi_0
evo=1.31.1=pypi_0
exceptiongroup=1.2.2=pypi_0
executing=2.2.0=pypi_0
fastjsonschema=2.21.1=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.18.0=pypi_0
fonttools=4.56.0=pypi_0
fqdn=1.5.1=pypi_0
freetype=2.12.1=h4a9f257_0
gdown=5.2.0=pypi_0
gmp=6.3.0=h6a678d5_0
gnutls=3.6.15=he1e5248_0
grpcio=1.71.0=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.6.1=pypi_0
importlib-resources=6.5.2=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.5=pypi_0
ipython=8.18.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.1.5=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.6=pypi_0
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_3
json5=0.10.0=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter-client=7.4.9=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter-events=0.12.0=pypi_0
jupyter-packaging=0.12.3=pypi_0
jupyter-server=2.15.0=pypi_0
jupyter-server-fileid=0.9.3=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyter-server-ydoc=0.8.0=pypi_0
jupyter-ydoc=0.2.5=pypi_0
jupyterlab=3.6.8=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.3=pypi_0
jupyterlab-widgets=3.0.13=pypi_0
kiwisolver=1.4.7=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.16=hb9589c4_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h6a678d5_0
libblas=3.9.0=1_h86c2bf4_netlib
libcblas=3.9.0=5_h92ddd45_netlib
libdeflate=1.22=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=13.2.0=h69a702a_0
libgfortran5=13.2.0=ha4646dd_0
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
liblapack=3.9.0=5_h92ddd45_netlib
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=hffd6297_1
libunistring=0.9.10=h27cfd23_0
libuv=1.48.0=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_1
lietorch=0.2=pypi_0
lz4=4.4.3=pypi_0
lz4-c=1.9.4=h6a678d5_1
markdown=3.7=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.9.4=pypi_0
matplotlib-inline=0.1.7=pypi_0
metis=5.1.0=h58526e2_1006
mistune=3.1.3=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_2
mkl_fft=1.3.11=py39h5eee18b_0
mkl_random=1.2.8=py39h1128e8f_0
mpfr=4.1.0=h9202a9a_1
natsort=8.4.0=pypi_0
nbclassic=1.2.0=pypi_0
nbclient=0.10.2=pypi_0
nbconvert=7.16.6=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
nettle=3.7.3=hbbd107a_1
notebook=6.5.7=pypi_0
notebook-shim=0.2.4=pypi_0
numexpr=2.10.2=pypi_0
numpy=1.23.0=pypi_0
numpy-quaternion=2023.0.4=pypi_0
open3d=0.15.2=pypi_0
opencv-python=4.7.0.72=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openssl=3.0.16=h5eee18b_0
overrides=7.7.0=pypi_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
pexpect=4.9.0=pypi_0
pillow=11.1.0=py39hcea889d_0
pip=25.0=py39h06a4308_0
platformdirs=4.3.7=pypi_0
prometheus-client=0.21.1=pypi_0
prompt-toolkit=3.0.50=pypi_0
protobuf=6.30.2=pypi_0
psutil=7.0.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
pycparser=2.22=pypi_0
pygments=2.19.1=pypi_0
pyparsing=3.2.3=pypi_0
pyquaternion=0.9.9=pypi_0
pysocks=1.7.1=pypi_0
python=3.9.21=he870216_1
python-dateutil=2.9.0.post0=pypi_0
python-json-logger=3.3.0=pypi_0
pytorch=1.10.2=py3.9_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2025.2=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=26.3.0=pypi_0
readline=8.2=h5eee18b_0
referencing=0.36.2=pypi_0
requests=2.32.3=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rosbags=0.9.23=pypi_0
rpds-py=0.24.0=pypi_0
ruamel-yaml=0.18.10=pypi_0
ruamel-yaml-clib=0.2.12=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.13.1=pypi_0
seaborn=0.13.2=pypi_0
send2trash=1.8.3=pypi_0
setuptools=75.8.0=py39h06a4308_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
sqlite=3.45.3=h5eee18b_0
stack-data=0.6.3=pypi_0
suitesparse=5.10.1=h9e50725_1
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.19.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
terminado=0.18.1=pypi_0
threadpoolctl=3.6.0=pypi_0
tinycss2=1.4.0=pypi_0
tk=8.6.14=h39e8969_0
tomli=2.2.1=pypi_0
tomlkit=0.13.2=pypi_0
torch-scatter=2.0.9=pypi_0
torchaudio=0.10.2=py39_cu113
torchvision=0.11.3=py39_cu113
tornado=6.4.2=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
types-python-dateutil=2.9.0.20241206=pypi_0
typing_extensions=4.12.2=py39h06a4308_0
tzdata=2025.2=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.3.0=pypi_0
wcwidth=0.2.13=pypi_0
webcolors=24.11.1=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.45.1=py39h06a4308_0
widgetsnbextension=4.0.13=pypi_0
xz=5.6.4=h5eee18b_1
y-py=0.6.2=pypi_0
ypy-websocket=0.8.4=pypi_0
zipp=3.21.0=pypi_0
zlib=1.2.13=h5eee18b_1
zstandard=0.23.0=pypi_0
zstd=1.5.6=hc292b87_0
